logging added to errors and successes error handling fixed added initialization to be optional with console argumnet

+++ src/beerbase/server.py
@@ -7,8 +7,9 @@ from flask import render_template
 from argparse import ArgumentParser, Namespace
 from pathlib import Path
 from logging import INFO, getLogger, Formatter, StreamHandler
+from typing import Optional
 
-from database import Database, Database_handler
+from database import Database, DatabaseHandler
 from datatypes import Base
 
 app = connexion.App("beerbase")
@@ -39,34 +40,53 @@ def parse_args() -> Namespace:
     parser.add_argument("--templates", help="Absolute path of templates folder", type=Path, required=True)
     parser.add_argument("--assets", help="Absolute path of assets folder", type=Path, required=True)
     parser.add_argument("--db", help="Absolute path of database file", type=Path, required=True)
-    parser.add_argument("--data", help="Absolute path of data file", type=Path, required=True)
+    parser.add_argument("--data", help="Absolute path of data file", type=Path, required=False)
+    parser.add_argument('--init', help="Initialize all tables and create them from 0", action='store_true')
+    parser.set_defaults(init=False)
 
     return parser.parse_args()
 
 
-def init_db(db_path: Path, csv_path: Path) -> None:
-    """Set's up database and loads contents of csv in.
+def init_db(db_path: Path, csv_path: Optional[Path] = None, init_tables: Optional[bool] = False) -> None:
+    """Sets up database and loads contents of csv in.
 
     Args:
         db_path (Path): Path to database file.
-        csv_path (Path): Path to csv datafile.
+        csv_path (Path): Path to csv datafile. Only imports data from it if init_tables is True.
+        init_tables (bool): Wether to drop all tables and create them again from 0.
     """
+    logger = getLogger("logger")
 
     if not db_path.exists():
-        file = db_path.open('w')
-        file.close()
+        try:
+            file = db_path.open('w')
+            file.close()
+        except Exception as exc:
+            logger.error(f'File could not be opened.\n'
+                         f'Cause:\n'
+                         f'{exc}')
+        else:
+            logger.info(f'File "{db_path}" created.')
 
-    db = Database(db_path=str(db_path))
-    Base.metadata.drop_all(db.engine)
-    Base.metadata.create_all(db.engine)
+    try:
+        db = Database(db_path=str(db_path))
 
-    if csv_path.exists():
-        handler = Database_handler()
-        handler.load_csv(csv_path)
+        if init_tables:
+            try:
+                Base.metadata.drop_all(db.engine)
+                Base.metadata.create_all(db.engine)
+            except Exception as exc:
+                logger.error(exc)
+
+            if csv_path.exists():
+                handler = DatabaseHandler()
+                handler.load_csv(csv_path)
+    except Exception as exc:
+        logger.error(exc)
 
 
 def set_logger() -> None:
-    """Set's up logger"""
+    """Sets up logger"""
 
     FORMAT = '%(levelname)-7s:%(asctime)s: %(message)s'
     formatter = Formatter(FORMAT)
@@ -86,7 +106,7 @@ def main():
     args = parse_args()
 
     # Initialize database
-    init_db(db_path=args.db, csv_path=args.data)
+    init_db(db_path=args.db, csv_path=args.data, init_tables=args.init)
 
     # Create app and link external file locations
     app.specification_dir = args.assets
